extension-py/requirements.txt
websocket-client==1.8.0
pyserial==3.4